Updated package.json.
Major version bump. All previous versions and interfaces are broken due to
th fact that everything is now TS+Jade instead of html and JavaScript.

On top of that implemented TypeScript 1.5 and EcmaScript 6 features.

Running on fumes here... Works again. Potentially will move all the
YHBT-elements out of the application itself and publish them separately
online in bower or something. Just for the kicks of it. Even though they
are not really reusable by others, they sure are usable by my project :D